# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  middleware.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  middleware.SwaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorInfo'
      success:
        example: false
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@travellink.local
    name: API Support
    url: http://www.swagger.io/support
  description: Travel Linked List API - Browse, create, and share travel plans
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TravelLink API
  version: "1.0"
paths:
  /admin/promotions/{id}/approve:
    patch:
      consumes:
      - application/json
      description: Admin can approve a promotion request. If no plan is specified,
        user role is upgraded to traveller. (admin only)
      parameters:
      - description: Promotion Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Admin notes (max 500 chars)
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Promotion request approved
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Promotion request not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Approve a promotion request
      tags:
      - admin
  /admin/promotions/{id}/reject:
    patch:
      consumes:
      - application/json
      description: Admin can reject a promotion request with optional notes. (admin
        only)
      parameters:
      - description: Promotion Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Admin notes (max 500 chars)
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Promotion request rejected
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Promotion request not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Reject a promotion request
      tags:
      - admin
...
  /nodes/{id}:
    delete:
      description: Admin can delete a node (admin only)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Delete a node
      tags:
      - admin
    get:
      description: Retrieve detailed information about a node (attraction or transition).
        Details include full information for the specific node type. (public endpoint)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      summary: Get node details
      tags:
      - nodes
  /nodes/{id}/approve:
    patch:
      description: Admin can approve a user-created node for inclusion in travel plans
        (admin only)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node approved successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Approve a user-created node
      tags:
      - admin
  /nodes/{id}/disapprove:
    patch:
      description: Admin can disapprove a node, removing it from travel plans (admin
        only)
      parameters:
      - description: Node ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Node disapproved successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Node not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Disapprove a node
      tags:
      - admin
...
  /plans/{id}:
    delete:
      description: Admin can soft-delete a travel plan (admin only)
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Plan deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Delete a travel plan
      tags:
      - admin
    get:
      description: Retrieve detailed information about a published travel plan including
        nodes (public endpoint)
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Plan details with nodes
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      tags:
      - plans
...
  /plans/{id}/suspend:
    patch:
      description: Admin can suspend a travel plan (admin only)
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Plan suspended successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Plan not found
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Suspend a travel plan
      tags:
      - admin
...
  /users/{id}/deactivate:
    patch:
      description: Admin can soft-delete (deactivate) a user account. Admin cannot
        deactivate self. (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deactivated successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "403":
          description: Cannot deactivate own account
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Deactivate user account
      tags:
      - admin
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Admin can change a user's role (simple/traveller/admin). Admin
        cannot change admin's own role. (admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Role change request (role: simple|traveller|admin)'
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: User role updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid role or cannot change own role
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "403":
          description: Cannot change own role
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.SwaggerErrorResponse'
      security:
      - Bearer: []
      summary: Change user role
      tags:
      - admin
securityDefinitions:
  Bearer:
    description: 'Enter JWT token in the format: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"